+++ .iex.exs
@@ -1,0 +1,2 @@
+Repatch.setup(enable_global: true)
+require Repatch